+++ Sources/ZcashLightClientKit/Error/ZcashErrorCode.swift
@@ -177,6 +177,10 @@ public enum ZcashErrorCode: String {
     case rustGetWalletSummary = "ZRUST0056"
     /// Error from rust layer when calling ZcashRustBackend.
     case rustProposeTransferFromURI = "ZRUST0057"
+    /// Error from rust layer when calling ZcashRustBackend.
+    case rustListAccounts = "ZRUST0058"
+    /// Error from rust layer when calling ZcashRustBackend.rustIsSeedRelevantToAnyDerivedAccount
+    case rustIsSeedRelevantToAnyDerivedAccount = "ZRUST0059"
     /// SQLite query failed when fetching all accounts from the database.
     case accountDAOGetAll = "ZADAO0001"
     /// Fetched accounts from SQLite but can't decode them.