+++ CHANGELOG.md
@@ -18,6 +18,29 @@ Possible errors:
 - `ZcashError.rustProposeTransferFromURI`
 - Other errors that `sentToAddress` can throw
 
+## Removed
+
+- `SDKSynchronizer.latestUTXOs`
+
+## Checkpoints
+
+Mainnet
+
+````
+Sources/ZcashLightClientKit/Resources/checkpoints/mainnet/2430000.json
+...
+Sources/ZcashLightClientKit/Resources/checkpoints/mainnet/2447500.json
+````
+
+
+Testnet
+
+````
+Sources/ZcashLightClientKit/Resources/checkpoints/testnet/2750000.json
+...
+Sources/ZcashLightClientKit/Resources/checkpoints/testnet/2770000.json
+````
+
 # 2.0.11 - 2024-03-08
 
 ## Changed